fully working script with some additional refactoring
