+++ Week 2/Day 2/Submissions/Aleksandr.Medvedev/script.py
@@ -14,6 +14,9 @@ from sklearn.dummy import DummyClassifier
 from sklearn.svm import SVC
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.neighbors import KNeighborsClassifier
+from sklearn.linear_model import LogisticRegression
+
+from brand import Brand, Burberry, Versace, DolceGabbana
 
 
 class Model:
@@ -33,7 +36,7 @@ class Model:
             'KNeighbors': KNeighborsClassifier
         }
         
-        self.clf = classifiers(self.model_type)(**clf_params)
+        self.clf = classifiers[self.model_type](**clf_params)
         
     def fit_and_score(self, X_train, X_test, y_train, y_test):
         self.clf.fit(X_train, y_train)
@@ -46,7 +49,7 @@ def encode_words(data):
     integer_encoded = label_encoder.fit_transform(data['words'])
     
     # binary encode
-    onehot_encoder = OneHotEncoder(sparse=False)
+    onehot_encoder = OneHotEncoder(sparse=False, categories='auto')
     integer_encoded = integer_encoded.reshape(len(integer_encoded), 1)
     onehot_encoded = onehot_encoder.fit_transform(integer_encoded)
     return onehot_encoded
@@ -54,112 +57,9 @@ def encode_words(data):
 
 def encode_brands(data):
     label_encoder = LabelEncoder()
-    integer_encoded = label_encoder.fit_transform(data['brands'])
+    integer_encoded = label_encoder.fit_transform(data['brand'])
     return integer_encoded
 
-
-
-    
-
-
-class Brand:
-    def __init__(self, name, urls, relevance_threshold):
-        
-        supported_brands = ['burberry', 'versace', 'd&g']
-        if name not in supported_brands:
-            raise ValueError(f'unsupported brand {name}')
-        
-        self.name = name
-        self.urls = urls
-        self.relevance_threshold = relevance_threshold
-    
-    def scrap_links(self):
-        links = []
-        for url in self.urls:
-            r = requests.get(url)
-            html_doc = r.text
-            soup = BeautifulSoup(html_doc)
-
-            for link in soup.find_all("a"):
-                l = self.link_selector(link) 
-                if l is not None:
-                    doc.append(l)
-
-        # DEDUPLICATING THE LIST OF LINKS
-        unique_links = set(links)
-        return unique_links
-    
-    def links_to_dataframe(self, links):
-        # CREATING A DICTIONARY WITH WORDS : COUNTS AND KEY : VALUE PAIRS
-        result = {}
-        for link in links:
-            words = self.word_extractor(link)
-            for word in words:
-                if word in result:
-                    result[word] += 1
-                else:
-                    result[word] = 1
-
-        words = list(result.keys())
-        counts = list(result.values())
-
-        # TURNING THE DICTIONARY INTO A DATAFRAME, SORTING & SELECTING FOR RELEVANCE
-        links_frame = pd.DataFrame.from_dict({
-            "words": words,
-            "counts": counts,
-        })
-
-        sorted_links = links_frame.sort_values("counts", ascending = True)
-        relevant_links = sorted_links[sorted_links['counts'] > relevance_threshold]
-        relevant_links.loc[:, 'brand'] = self.name
-        return relevant_links
-    
-    def plot_most_used(self, data, image_path):
-        plt.barh(words, counts, color = "#C19A6B")
-        plt.title(f'{self.name} most frequent words')
-        plt.xticks(np.arange(0, 18, step=2))
-        plt.savefig(image_path)
-    
-
-class Burberry(Brand):
-    
-    def link_selector(self, link):
-        l = link.get("href")
-        if "-p80" in l: 
-            return l
-        else:
-            return None
-            
-    def word_extractor(self, link):
-        return link.replace("/", "").split("-")
-
-
-class Versace(Brand):
-    
-    def link_selector(self, link):
-        a = t.get("href")
-        if a.startswith("/us/en-us/women/new-arrivals/new-in/") and not link.startswith("/us/en-us/women/new-arrivals/new-in/?"):
-            return a
-        else:
-            return None
-        
-    def word_extractor(self, link):
-        words = link.replace("/us/en-us/women/new-arrivals/new-in/", "") .split("/")
-        words = words[0].split("-")
-        return words
-        
-        
-class DolceGabbana(Brand):
-    
-    def link_selector(self, link):
-        a = t.get("aria-label")
-        if a != None and a.startswith("Visit"):
-            return a
-        else:
-            return None
-            
-    def word_extractor(self, link):
-        return link.replace("Visit", "").replace(" product page","").split(" ")                                                                                                                                      
 
 if __name__ == '__main__':
     
@@ -175,19 +75,24 @@ if __name__ == '__main__':
         config = json.load(config_file)
        
     
-    burberry = Burberry('burberry', config['brands']['burberry']['urls'], config['brands']['burberry']['relevance_threshold'])
-    versace = Versace('versace', config['brands']['versace']['urls'], config['brands']['versace']['relevance_threshold'])
-    dolce_gabbana = DolceGabbana('d&g', config['brands']['d&g']['urls'], config['brands']['d&g']['relevance_threshold'])
+    burberry_conf = config['brands']['burberry']
+    burberry = Burberry('burberry', burberry_conf['urls'], burberry_conf['relevance_threshold'])
+
+    versace_conf = config['brands']['versace']
+    versace = Versace('versace', versace_conf['urls'], versace_conf['relevance_threshold'])
+
+    dg_conf = config['brands']['d&g']
+    dolce_gabbana = DolceGabbana('d&g', dg_conf['urls'], dg_conf['relevance_threshold'])
     
     brands = [burberry, versace, dolce_gabbana]
     data = []
     for brand in brands:
         links = brand.scrap_links()
-        if verbose:
-            print(f'for brand {brand} we found {len(links)} unique links')
+        if args.verbose:
+            print(f'for brand {brand.name} we found {len(links)} unique links')
         brand_data = brand.links_to_dataframe(links)
         data.append(brand_data)
-        brand.plot(brand_data, f'{brand_name}.png')
+        brand.plot_most_used(brand_data, f'{brand.name}.png')
         
     data = pd.concat(data, axis=0, ignore_index=True)
     words_encoded = encode_words(data)
@@ -196,13 +101,14 @@ if __name__ == '__main__':
     
     X_train, X_test, y_train, y_test = train_test_split(words_encoded, brands_encoded, test_size=0.33, random_state=args.random_state)
     
-    accuracies = {}
+    accuracies = {'model': [], 'accuracy': []}
     for model_type, params_list in config['models'].items():
         for params in params_list:
             model = Model(model_type, **params)
             accuracy = model.fit_and_score(X_train, X_test, y_train, y_test)
-            accuracies[model.name] = accuracy
-            if verbose:
+            accuracies['model'].append(model.name)
+            accuracies['accuracy'].append(accuracy)
+            if args.verbose:
                 print(f'for model {model.name} we got {accuracy:.3f} accuracy')
                       
     results = pd.DataFrame.from_dict(accuracies)
